fix: critical data-integrity, sync, and checkout fixes from audit
- Fractional stock now deducts the exact amount (was truncated to int) by
  threading deltaDecimal through StockMovementService + recordSaleDeductions
- Server pull no longer overwrites unsynced local product/variant edits
- AI-created sales now deduct inventory, atomically with the sale insert
  (previously they never deducted stock)
- Logout flushes pending data first and skips the local wipe while records
  are still unsynced, preventing loss of offline sales
- POS sale + inventory deduction are now atomic via CheckoutService
- Sync no longer infinitely re-queues a missing parent for orphaned
  inventory levels; surfaces them as failures instead
- Bottom nav item no longer overflows under larger text scale
- Add regression tests: fractional stock, pull guard, checkout atomicity

Note: sync_service.dart also carries a small pre-existing local FK self-heal
hunk for variants that was already in the working tree.

+++ lib/features/inventory/domain/repositories/i_inventory_repository.dart
@@ -16,8 +16,10 @@ abstract class IInventoryRepository {
     String? note,
   });
 
+  // qty is a double so weight/fraction products (e.g. 0.5 kg) deduct the exact
+  // amount sold instead of being rounded to a whole unit.
   Future<void> recordSaleDeductions({
-    required List<({String variantId, int qty})> items,
+    required List<({String variantId, double qty})> items,
     required String businessId,
     required String? branchId,
   });
